Option _compressBitmaps_ enables compression.
The option _compressBitmaps_ refers to _jsPDF.addImage(compression)_.
During the conversion svg2pdf uses _/nodes/image.ts_ to render images.
Those are transformed with _ImageNode.renderCore()_ in _/nodes/image.ts_.
The option _compressBitmaps_ is passed to _ImageNode.renderCore()_ in order to enable the available compression of jsPDF.
Values are of type ImageCompression.